move store recorder to 'mock/mockstore'
