+++ pkg/wait/wait.go
@@ -19,8 +19,6 @@ package wait
 import (
 	"log"
 	"sync"
-
-	"github.com/coreos/etcd/pkg/testutil"
 )
 
 type Wait interface {
@@ -59,29 +57,6 @@ func (w *List) Trigger(id uint64, x interface{}) {
 		ch <- x
 		close(ch)
 	}
-}
-
-type WaitRecorder struct {
-	Wait
-	testutil.Recorder
-}
-
-type waitRecorder struct {
-	testutil.RecorderBuffered
-}
-
-func NewRecorder() *WaitRecorder {
-	wr := &waitRecorder{}
-	return &WaitRecorder{Wait: wr, Recorder: wr}
-}
-func NewNop() Wait { return NewRecorder() }
-
-func (w *waitRecorder) Register(id uint64) <-chan interface{} {
-	w.Record(testutil.Action{Name: "Register"})
-	return nil
-}
-func (w *waitRecorder) Trigger(id uint64, x interface{}) {
-	w.Record(testutil.Action{Name: "Trigger"})
 }
 
 type waitWithResponse struct {
